Finalize Channel Switch
