+++ app/actions/views/channel.test.js
@@ -116,14 +116,19 @@ describe('Actions.Views.Channel', () => {
             },
             channels: {
                 currentChannelId,
+                channels: {
+                    'channel-id': {id: 'channel-id', display_name: 'Test Channel'},
+                },
+                myMembers: {
+                    'channel-id': {channel_id: 'channel-id', user_id: currentUserId, mention_count: 0, msg_count: 0},
+                },
             },
             teams: {
+                currentTeamId,
                 teams: {
-                    currentTeamId,
-                    currentTeams: {
-                        [currentTeamId]: {
-                            name: currentTeamName,
-                        },
+                    [currentTeamId]: {
+                        id: currentTeamId,
+                        name: currentTeamName,
                     },
                 },
             },
@@ -147,14 +152,13 @@ describe('Actions.Views.Channel', () => {
         const receivedChannel = storeActions.some((action) => action.type === MOCK_RECEIVE_CHANNEL_TYPE);
         expect(receivedChannel).toBe(true);
 
-        const storeBatchActions = storeActions.filter(({type}) => type === 'BATCHING_REDUCER.BATCH');
-        const selectedChannel = storeBatchActions[0].payload.some((action) => action.type === MOCK_SELECT_CHANNEL_TYPE);
+        const selectedChannel = storeActions.some(({type}) => type === MOCK_RECEIVE_CHANNEL_TYPE);
         expect(selectedChannel).toBe(true);
     });
 
     test('handleSelectChannelByName failure from null currentTeamName', async () => {
         const failStoreObj = {...storeObj};
-        failStoreObj.entities.teams.teams.currentTeamId = 'not-in-current-teams';
+        failStoreObj.entities.teams.currentTeamId = 'not-in-current-teams';
         store = mockStore(failStoreObj);
 
         await store.dispatch(handleSelectChannelByName(currentChannelName, null));
@@ -168,6 +172,7 @@ describe('Actions.Views.Channel', () => {
     });
 
     test('handleSelectChannelByName failure from no permission to channel', async () => {
+        store = mockStore({...storeObj});
         actions.getChannelByNameAndTeamName = jest.fn(() => {
             return {
                 type: 'MOCK_ERROR',
@@ -283,12 +288,13 @@ describe('Actions.Views.Channel', () => {
         [`not-${currentChannelId}`, false],
     ];
     test.each(handleSelectChannelCases)('handleSelectChannel dispatches selectChannelWithMember', async (channelId, fromPushNotification) => {
-        store = mockStore({...storeObj});
+        const testObj = {...storeObj};
+        testObj.entities.teams.currentTeamId = currentTeamId;
+        store = mockStore(testObj);
 
         await store.dispatch(handleSelectChannel(channelId, fromPushNotification));
         const storeActions = store.getActions();
-        const storeBatchActions = storeActions.find(({type}) => type === 'BATCHING_REDUCER.BATCH');
-        const selectChannelWithMember = storeBatchActions.payload.find(({type}) => type === ChannelTypes.SELECT_CHANNEL);
+        const selectChannelWithMember = storeActions.find(({type}) => type === ChannelTypes.SELECT_CHANNEL);
         const viewedAction = storeActions.find(({type}) => type === MOCK_CHANNEL_MARK_AS_VIEWED);
         const readAction = storeActions.find(({type}) => type === MOCK_CHANNEL_MARK_AS_READ);
 
@@ -297,16 +303,23 @@ describe('Actions.Views.Channel', () => {
             data: channelId,
             extra: {
                 channel: {
-                    data: channelId,
+                    id: channelId,
+                    display_name: 'Test Channel',
                 },
                 member: {
-                    data: {
-                        member: {},
-                    },
+                    channel_id: channelId,
+                    user_id: currentUserId,
+                    mention_count: 0,
+                    msg_count: 0,
                 },
+                teamId: currentTeamId,
             },
         };
-        expect(selectChannelWithMember).toStrictEqual(expectedSelectChannelWithMember);
+        if (channelId.includes('not')) {
+            expect(selectChannelWithMember).toBe(undefined);
+        } else {
+            expect(selectChannelWithMember).toStrictEqual(expectedSelectChannelWithMember);
+        }
         expect(viewedAction).not.toBe(null);
         expect(readAction).not.toBe(null);
     });
