+++ app/components/post_list/post_list.js
@@ -24,7 +24,7 @@ import {t} from 'app/utils/i18n';
 import DateHeader from './date_header';
 import NewMessagesDivider from './new_messages_divider';
 
-const INITIAL_BATCH_TO_RENDER = 15;
+const INITIAL_BATCH_TO_RENDER = 7;
 const SCROLL_UP_MULTIPLIER = 3.5;
 const SCROLL_POSITION_CONFIG = {
 
@@ -129,7 +129,7 @@ export default class PostList extends PureComponent {
             this.shouldScrollToBottom = false;
         }
 
-        if (!this.hasDoneInitialScroll && this.props.initialIndex > 0 && this.state.contentHeight) {
+        if (!this.hasDoneInitialScroll && this.props.initialIndex > 0 && this.state.contentHeight > 53) {
             this.scrollToInitialIndexIfNeeded(this.props.initialIndex);
         }
 
@@ -169,7 +169,7 @@ export default class PostList extends PureComponent {
     handleContentSizeChange = (contentWidth, contentHeight) => {
         if (this.state.contentHeight !== contentHeight) {
             this.setState({contentHeight}, () => {
-                if (this.state.postListHeight && contentHeight < this.state.postListHeight && this.props.extraData) {
+                if (this.state.postListHeight && contentHeight < this.state.postListHeight && this.props.extraData && contentHeight > 53) {
                     // We still have less than 1 screen of posts loaded with more to get, so load more
                     this.props.onLoadMoreUp();
                 }
@@ -392,11 +392,13 @@ export default class PostList extends PureComponent {
     };
 
     flatListScrollToIndex = (index) => {
-        this.flatListRef.current.scrollToIndex({
-            animated: false,
-            index,
-            viewOffset: 0,
-            viewPosition: 1, // 0 is at bottom
+        this.animationFrameInitialIndex = requestAnimationFrame(() => {
+            this.flatListRef.current.scrollToIndex({
+                animated: false,
+                index,
+                viewOffset: 0,
+                viewPosition: 1, // 0 is at bottom
+            });
         });
     }
 
@@ -486,7 +488,6 @@ export default class PostList extends PureComponent {
                 keyExtractor={this.keyExtractor}
                 ListFooterComponent={this.props.renderFooter}
                 maintainVisibleContentPosition={SCROLL_POSITION_CONFIG}
-                maxToRenderPerBatch={INITIAL_BATCH_TO_RENDER + 1}
                 onContentSizeChange={this.handleContentSizeChange}
                 onLayout={this.handleLayout}
                 onScroll={this.handleScroll}
@@ -496,6 +497,7 @@ export default class PostList extends PureComponent {
                 scrollEventThrottle={60}
                 refreshControl={refreshControl}
                 nativeID={scrollViewNativeID}
+                windowSize={50}
             />
         );
     }
