+++ app/constants/view.js
@@ -69,7 +69,6 @@ const ViewTypes = keyMirror({
     SAML: null,
 
     SET_INITIAL_POST_VISIBILITY: null,
-    INCREASE_POST_VISIBILITY: null,
     RECEIVED_FOCUSED_POST: null,
     LOADING_POSTS: null,
     SET_LOAD_MORE_POSTS_VISIBLE: null,
