+++ app/store/store.js
@@ -62,7 +62,7 @@ export default function configureAppStore(initialState) {
         ['typing'],
     );
 
-    const channelViewBlackList = {loading: true, refreshing: true, loadingPosts: true, postVisibility: true, retryFailed: true, loadMorePostsVisible: true};
+    const channelViewBlackList = {loading: true, refreshing: true, loadingPosts: true, retryFailed: true, loadMorePostsVisible: true};
     const channelViewBlackListFilter = createTransform(
         (inboundState) => {
             const channel = {};
