+++ babel.config.js
@@ -2,12 +2,11 @@
 // See LICENSE.txt for license information.
 module.exports = {
     presets: ['module:metro-react-native-babel-preset'],
-
-    // env: {
-    //     production: {
-    //         plugins: ['transform-remove-console'],
-    //     },
-    // },
+    env: {
+        production: {
+            plugins: ['transform-remove-console'],
+        },
+    },
     plugins: [
         '@babel/plugin-transform-runtime',
         ['module-resolver', {
