+++ package.json
@@ -24,7 +24,7 @@
     "intl": "1.2.5",
     "jail-monkey": "2.3.1",
     "jsc-android": "241213.2.0",
-    "mattermost-redux": "github:mattermost/mattermost-redux#af7a071de69e778a633ab182367568c21977f6b0",
+    "mattermost-redux": "github:mattermost/mattermost-redux#0a5ec3fa0f816ad8bac870520f5fd6a8436ee4ed",
     "mime-db": "1.43.0",
     "moment-timezone": "0.5.27",
     "prop-types": "15.7.2",
