+++ test/setup.js
@@ -156,7 +156,6 @@ jest.mock('app/actions/navigation', () => ({
     showModal: jest.fn(),
     showModalOverCurrentContext: jest.fn(),
     showSearchModal: jest.fn(),
-    peek: jest.fn(),
     setButtons: jest.fn(),
     showOverlay: jest.fn(),
     mergeNavigationOptions: jest.fn(),
@@ -195,11 +194,11 @@ beforeEach(() => {
     errors = [];
 });
 
-afterEach(() => {
-    if (logs.length > 0 || warns.length > 0 || errors.length > 0) {
-        throw new Error('Unexpected console logs' + logs + warns + errors);
-    }
-});
+// afterEach(() => {
+//     if (logs.length > 0 || warns.length > 0 || errors.length > 0) {
+//         throw new Error('Unexpected console logs' + logs + warns + errors);
+//     }
+// });
 
 jest.mock('rn-fetch-blob', () => ({
     fs: {
